buffer: move buffer::raw to new header
this allows new raw buffer types to live outside of buffer.cc, and
potentially outside of libcommon entirely
